\name{tartness}
\alias{tartness}
\title{Tartness Rankings of Salad Dressings}
\description{
  The data were collected by Vargo (1989). Each of 32 judges is asked to rank
  four salad dressing preparations according to tartness, with a rank of 1
  being assigned to the formulation judged to be the most tart.
}
\usage{data(tartness)}
\format{a data frame consisting the rankings and their frequencies.}
\source{
  Critchlow, D.E., & Fligner, M.A. (1991). Paired comparison, triple
  comparison, and ranking experiments as generalized linear models, and their
  implementation in GLIM.
  \emph{Psychometrika}, \bold{56}, 517--533.
}
\references{
  Vargo, M.D. (1989). Microbiological spoilage of a moderate acid food system
  using a dairy-based salad dressing model. Unpublished masters thesis, Ohio
  State University, Department of Food Science and Nutrition, Columbus, OH.
}
\examples{
data(tartness)
}
\keyword{datasets}
