\name{celebrities}
\alias{celebrities}
\title{Choice Among Celebrities}
\description{
  This data set provides the absolute choice frequencies of
  234 subjects choosing between pairs of nine celebrities.
  LBJ, HW, and CDG are politicians; JU, CY, and AJF are
  athletes; BB, ET, and SL are female movie stars.
  Row stimuli are chosen over column stimuli.
}
\usage{data(celebrities)}
\format{A square data frame with a diagonal of zeros.}
\references{
  Rumelhart, D.L., & Greeno, J.G. (1971).
  Similarity between stimuli: An experimental test of the
  Luce and Restle choice models.
  \emph{Journal of Mathematical Psychology}, \bold{8}, 370-381.
}
\keyword{datasets}
