\name{pork}
\alias{pork}
\title{Pork Tasting Data}
\description{
  This data set provides the individual choice matrices
  of two judges choosing between pairs of three samples
  of pork meet. Each judge does five repetitions. The
  data are stored in a 3d array, the first five matrices
  of which correspond to the five repetitions of the first
  judge, the last five to the repetitions of the second
  judge. Row stimuli are chosen over column stimuli.
}
\usage{data(pork)}
\format{A 3d array consisting of ten square matrices.}
\references{
  Bradley, R.A., & Terry, M.E. (1952).
  Rank analysis of incomplete block designs.
  I. The method of paired comparisons.
  \emph{Biometrika}, \bold{39}, 324-345.
}
\keyword{datasets}
